Added PrimaryAssembly attribute to NAnt task
Because order within a FileSet is not maintained in NAnt, listing the primary assembly first is insufficient.  This now adds an element "<primaryassembly>" to the ILMerge task that accepts one and only one primary assembly.  All other assemblies to be merged should be placed in the "<assemblies>" element.

Also modified properties match new issue tracker.

Fixes Defect #1